backend/family-os-mvp-api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
openai==1.58.1
python-dotenv==1.0.1
pytest==8.3.4
